+++ driver/inc/Driver_Protocol.h
@@ -77,14 +77,14 @@ typedef struct {
     // 协议定义
     uint16_t id;     // 编号
     uint16_t length; // 数据段长度
-    // 存储相关
-    uint16_t offset; // 寄存器偏移
     // 接收相关
     uint16_t receiving;    // 是否接收并解包该协议 @todo:未实现
     uint16_t receiveCount; // 接收计数 @todo:未实现
     // 发送相关
     float         frequency;  // 发送频率 (实际发送频率会略低一些) todo:未实现
     TaskHandle_t *taskHandle; // 发送任务 @todo:未实现
+    // 存储相关
+    uint16_t offset; // 寄存器偏移
 } ProtocolInfo_Type;
 
 void     Protocol_Get_Info(uint16_t id, uint16_t *offset, uint16_t *length);