+++ user/template/interrupt.c
@@ -94,7 +94,8 @@ void TIM2_IRQHandler(void) {
  */
 
 void NMI_Handler(void) {
-    printf("NMI_Handler");
+    while (1) {
+    }
 }
 
 /**
@@ -103,7 +104,8 @@ void NMI_Handler(void) {
  * @return None
  */
 void HardFault_Handler(void) {
-    printf("HardFault_Handler");
+    while (1) {
+    }
 }
 
 /**
@@ -112,7 +114,8 @@ void HardFault_Handler(void) {
  * @return None
  */
 void MemManage_Handler(void) {
-    printf("MemManage_Handler");
+    while (1) {
+    }
 }
 
 /**
@@ -121,7 +124,8 @@ void MemManage_Handler(void) {
  * @return None
  */
 void BusFault_Handler(void) {
-    printf("BusFault_Handler");
+    while (1) {
+    }
 }
 
 /**
@@ -130,7 +134,8 @@ void BusFault_Handler(void) {
  * @return None
  */
 void UsageFault_Handler(void) {
-    printf("UsageFault_Handler");
+    while (1) {
+    }
 }
 
 /**
@@ -139,7 +144,8 @@ void UsageFault_Handler(void) {
  * @return None
  */
 void DebugMon_Handler(void) {
-    printf("DebugMon_Handler");
+    while (1) {
+    }
 }
 
 // /**
@@ -148,7 +154,7 @@ void DebugMon_Handler(void) {
 //  * @return None
 //  */
 // void SVC_Handler(void) {
-//     //printf("SVC_Handler");
+//     //while(1){}
 // }
 
 // /**
@@ -157,7 +163,7 @@ void DebugMon_Handler(void) {
 //  * @return None
 //  */
 // void PendSV_Handler(void) {
-//     //printf("PendSV_Handler");
+//     //while(1){}
 // }
 
 // /**
@@ -166,5 +172,5 @@ void DebugMon_Handler(void) {
 //  * @return None
 //  */
 // void SysTick_Handler(void) {
-//     //printf("SysTick_Handler");
+//     //while(1){}
 // }